Moved ScopeContainer to support.scope package
